requirements.txt
ahrs
pyserial
dearpygui
numpy